# Improve allunique performance for strings, avoid computing length several times for iterators