--- a/test/sets.jl
+++ b/test/sets.jl
@@ -558,6 +558,11 @@ end
     @test allunique(x for x in [0.0, -0.0] if true)
     @test !allunique([NaN, NaN])
     @test !allunique(x for x in [NaN, NaN] if true)
+    # strings
+    @test !allunique("aaaa")
+    @test allunique("abcd")
+    @test !allunique("a" ^ 100) # longer than 32 but less than 1000
+    @test !allunique("a" ^ 10_000) # longer than 1000
     # ranges
     @test allunique(4:7)
     @test allunique(1:1)