--- a/test/sets.jl
+++ b/test/sets.jl
@@ -560,7 +560,9 @@ end
     @test !allunique(x for x in [NaN, NaN] if true)
     # strings
     @test !allunique("aaaa")
+    @test !allunique(@views("aaaa"[1:3])) # sub-string / abstractstring
     @test allunique("abcd")
+    @test allunique(@views("abcd"[1:3])) # sub-string / abstractstring
     @test !allunique("a" ^ 100) # longer than 32 but less than 1000
     @test !allunique("a" ^ 10_000) # longer than 1000
     # ranges