This PR fixes #50360 .

master: 
```julia
julia> VERSION
v"1.11.0-DEV.142"

julia> using Random

julia> using BenchmarkTools

julia> s=randstring(100000);

julia> @btime allunique($s)
  171.250 μs (4 allocations: 336 bytes)
false

julia> @btime allunique($s[1:100]) && allunique($s)
  258.552 ns (5 allocations: 456 bytes)
false
```

This PR:
```julia
julia> VERSION
v"1.11.0-DEV.142"

julia> using Random

julia> using BenchmarkTools

julia> s=randstring(100000);

julia> @btime allunique($s)
  72.362 ns (4 allocations: 336 bytes)
false

julia> @btime allunique($s[1:100]) && allunique($s)
  91.118 ns (5 allocations: 456 bytes)
false
```