@bvdmitri 2023-07-23
@oscardssmith There is failing test on CI:

```julia
Error in testset abstractarray:
--
  | Error During Test at /cache/build/default-amdci4-6/julialang/julia-master/julia-a96fdcc811/share/julia/test/abstractarray.jl:762
  | Test threw exception
  | Expression: #= /cache/build/default-amdci4-6/julialang/julia-master/julia-a96fdcc811/share/julia/test/abstractarray.jl:762 =# @inferred(cat3v(As)) == zeros(2, 2, 2)
  | return type Array{Float64, 3} does not match inferred return type Array{_A, 3} where _A
```

I'm not sure how is it related to the changes made in this PR. I tried to execute the test locally and it passed just in REPL.

```julia
julia> using Test

julia> As = [zeros(2, 2) for _ = 1:2]
2-element Vector{Matrix{Float64}}:
 [0.0 0.0; 0.0 0.0]
 [0.0 0.0; 0.0 0.0]

julia> @test @inferred(cat(As...; dims=Val(3))) == zeros(2, 2, 2)
Test Passed

julia> cat3v(As) = cat(As...; dims=Val(3))
cat3v (generic function with 1 method)

julia> @test @inferred(cat3v(As)) == zeros(2, 2, 2)
Test Passed

julia> @test @inferred(cat(As...; dims=Val((1,2)))) == zeros(4, 4)
Test Passed
```

I couldn't run the whole testset locally with `make testall` due to some problems with the fetching some data from the internet?

There is also a whitecheck failing, says there is a trailing whitespace. Is there a command to automatically fix all code-style related problems?