@LilithHafner 2023-08-03
> ```julia
>  return type Array{Float64, 3} does not match inferred return type Array{_A, 3} where _A
> ```

This is not your fault



> There is also a whitecheck failing, says there is a trailing whitespace. Is there a command to automatically fix all code-style related problems?

Not that I know of, but the check should also list the line number.