@LilithHafner 2023-08-07
> What is the current status of this PR?

My two biggest concerns are

- CI is failing (you can click on "details" next to the failed tests to see what exactly failed). It's true that we have pretty unreliable CI and oftentimes failures are not the fault of the PR author. However, we still need to look into test failures every time to confirm that they are not related. In this case, I believe this PR introduces some new failures.
- I am unaware of any code that calls `allunique(::AbstractString)` and can't think of any use case that warrants calling that method, and am hesitant to add code to optimize the performance of a method that I don't know is ever used.