@adienes 2025-12-24
I get the intention of speeding up `allunique` but I think the implementation could be improved a little; a variable like `len_or_ncodeunits` is a bit of a code smell to me. maybe the `len > 2` check can be moved into `allunique(::AbstractString)` ?