@DilumAluthge 2026-01-31
@bvdmitri Are you still interested in this PR? If so, can you rebase on the latest Julia master, and fix the merge conflicts? This will allow us to see what the current state of CI is, as discussed by @LilithHafner.